Amend button descriptions in game selection menu
This makes the ‘Control Stick: Select’, ‘B: Cancel’, and ‘A: Confirm’ descriptions appear while in the game selection menu, while still keeping just ‘B: Cancel’ when on the ‘Press Start’ screen.

To do this, this patches the function that updates the button descriptions, replacing two jump table entries, allowing our own code to be run (and determine which descriptions to display) while within the gameplay menu.