First pass at reading discs, load banner into menu
This creates a thread that handles disc access, including adapting to the user opening the lid, and either changing disc or having no disc inserted. This state then gets reported to the GameCube menu as per the stock IPL.

So far, it’ll just load enough of the disc to load the banner into the ‘game play’ menu.