Maintain a navigation stack for the gameplay menu
This makes it a bit simpler to maintain the different menu flows for the disc drive and FlippyDrive modes.

As before, pressing B while on the ‘Press Start’ screen will return to the top-level gameplay menu if the disc drive is active, and back to the ‘loader’ menu if the FlippyDrive is active – just now with less device-specific code.