Allow the boot error suppression to be configured
Previously, Cubeboot would always suppress error messages on boot relating to settings and the RTC. This gives the user option to leave these error messages in place, if they wish to be informed when their SRAM battery has lost power.

Additionally, this patches the calls that determines if an error has occurred, rather than the resulting menu state – this allows it to use the same code path as normal, setting the menu state depending on if `bs2tick` reports that the disc cover’s open.