Increase delay after entering bypass mode
Previously, on NTSC 1.0, the disc worker thread might briefly detect that the disc cover is open during startup. This would cause the automatic device switching logic to move onto the next boot device, or the GameCube to go to the menu.

We already have a delay after switching into bypass mode. Not my favourite kind of fix, but upping this delay from 10ms to 20ms seems to resolve the issue.